+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Add `debug` messages
   * Resolve relative paths at middleware setup
 
 1.1.6 / 2014-08-10